requirements.txt
# Core Data Processing
numpy
pandas

# API & Data Fetching
requests
yfinance
ta

# Database
psycopg2-binary
sqlalchemy
alembic

# DLT (Data Load Tool)
dlt[postgres]

# DBT (Data Build Tool)
dbt-core
dbt-postgres

# Machine Learning
pycaret
scikit-learn
scipy
xgboost

# MLflow (Experiment Tracking)
mlflow

# Configuration Management
python-dotenv
pydantic
pyyaml

# Orchestration & Scheduling
apscheduler
schedule
pytz

# Logging & Monitoring
python-json-logger
loguru

# Utilities
click
tqdm
ydata-profiling

# Testing
pytest
pytest-cov
pytest-asyncio

# Code Quality
black
flake8
isort
mypy
pylint

# Documentation
sphinx
sphinx-rtd-theme